Re-organize the region code so it doesn't crash on startup
I forgor that loading a class that extends another class which doesn't
exist crashes the game 💀💀💀
