+++ build.gradle
@@ -95,7 +95,7 @@ dependencies {
     minecraft "net.minecraftforge:forge:${config.MINECRAFT_VERSION}-${config.FORGE_VERSION}"
     annotationProcessor 'org.spongepowered:mixin:0.8.5:processor'
 
-    implementation fg.deobf("com.github.glitchfiend:TerraBlender-forge:${config.MINECRAFT_VERSION}-${config.TERRABLENDER_VERSION}")
+    compileOnly fg.deobf("com.github.glitchfiend:TerraBlender-forge:${config.MINECRAFT_VERSION}-${config.TERRABLENDER_VERSION}")
 }
 
 jar {
